refactor: :recycle: rename file to get_sprout_specific_resource_errors
